--- android/feature/home/src/main/kotlin/com/andlife/home/model/HomeUiModel.kt
@@ -1,3 +1,0 @@
-package com.andlife.home.model
-
-class HomeUiModel
