+++ android/feature/home/src/main/kotlin/com/andlife/home/model/home/HomeUiModel.kt
@@ -1,0 +1,3 @@
+package com.andlife.home.model.home
+
+class HomeUiModel
